ZoomH6: fix send command always returning false
after sending the desired command to the Zoom Recorder we wait for a
given delay time and the send a release button command.
the release button command will not always return status from the
recorder, so checking its reply and returning false if no reply might
cause us to always return false even if the command was successfully
committed.